// src/app/models/gestion-continuidad/GestionContinuidad.ts
export class GestionContinuidad {
    pomIdPrestOrdm: number;
    gcoIdCodigoEstado: any;
    gcoIdCodigoMotivo: number;
    gcoDirecPaciente: string;
    gcoRealizoAgendamiento: string;
    gcoObservaciones: string;
    pcaAgeCodigRecep: string;
    enviarCorreo: boolean;
    nombrePaciente: string;
    prePreDesc: string;
    prePreCodigo: string;
    serSerCodigo: string;
    serSerDesc: string;
}